+++ CHANGELOG.md
@@ -9,6 +9,8 @@ SMARTGEAR Project for Contao Open Source CMS
 - Smartgear : FileUsage bundle configuration updated
 - Smartgear : Adding `##useful_data##` & `##useful_data_filled##` tokens for forms' notifications
 - FormContact : Sender's address is now `##admin_email##` token instead of Smartgear's Owner' email address
+- StyleManager : the `rsce_listLogos` now has the same options as `rsce_listIcons`
+- StyleManager : the `headline` now has an `alignement` option
 
 1.0.20 - 2023-08-04
 - Smartgear : fixing files at `/templates`'s root being deleted when updating Smartgear.